Changed way we handle multiple forms on the external accounts view
Instead of having a different HTML form for each account form, we are
going to have 1 big form that contains the settings for each form.
Made a class that handles some of the complexity of this.  Instead of
a remove button, there will be a checkbox input that enables/disables
the account.

Refactored the AccountForm class a bit to make it work with
AccountFormset.

Still need to update the view/templates to work.

+++ apps/externalsites/forms.py
@@ -30,59 +30,38 @@ class AccountForm(forms.ModelForm):
     interface so that the view code doesn't need to special case each form
     """
 
-    label = NotImplemented
+    enabled = forms.BooleanField(required=False)
 
-    def __init__(self, team, data=None):
-        account = self.get_account(team)
-        self.allow_remove = (account is not None)
+    def __init__(self, team, data=None, **kwargs):
         self.team = team
-        forms.ModelForm.__init__(self, data, instance=account)
+        forms.ModelForm.__init__(self, data, **kwargs)
 
     @classmethod
-    def form_name(cls):
-        return str(cls)
-
-    def get_account(self, team):
-        ModelClass = self._meta.model
+    def get_account(cls, team):
+        ModelClass = cls._meta.model
         try:
             return ModelClass.objects.filter(team=team).get()
         except ModelClass.DoesNotExist:
             return None
 
-    @classmethod
-    def should_process_data(self, data):
-        return data.get('form-name') == self.form_name()
+    def delete_account(self):
+        if self.instance.id is not None:
+            self.instance.delete()
 
-    def save(self, commit=True):
+    def save(self):
         account = forms.ModelForm.save(self, commit=False)
         account.team = self.team
-        if commit:
-            account.save()
+        account.save()
         return account
 
-    def delete_account(self):
-        self.instance.delete()
-
 class KalturaAccountForm(AccountForm):
-    label = _("Kaltura")
+    partner_id = forms.IntegerField()
 
     class Meta:
         model = models.KalturaAccount
         fields = ['partner_id', 'secret']
 
-    def clean_partner_id(self):
-        partner_id = self.cleaned_data['partner_id']
-        if partner_id is not None:
-            try:
-                int(partner_id)
-            except ValueError:
-                raise forms.ValidationError(
-                    _('Partner ID must contain only numbers'))
-        return self.cleaned_data['partner_id']
-
 class BrightcoveAccountForm(AccountForm):
-    label = _("Brightcove")
-
     FEED_ALL_NEW = 'N'
     FEED_WITH_TAGS = 'T'
     FEED_CHOICES = (
@@ -115,8 +94,8 @@ class BrightcoveAccountForm(AccountForm):
                 ]
         return self.cleaned_data
 
-    def save(self, commit=True):
-        account = AccountForm.save(self, commit)
+    def save(self):
+        account = AccountForm.save(self)
         if self.cleaned_data['feed_enabled']:
             if self.cleaned_data['feed_type'] == self.FEED_ALL_NEW:
                 account.make_feed(self.cleaned_data['player_id'])
@@ -126,3 +105,80 @@ class BrightcoveAccountForm(AccountForm):
         else:
             account.remove_feed()
         return account
+
+class AccountFormset(object):
+    """dict-like object that contains multiple account forms.
+
+    For each form in form classes we will instatiate it with a unique prefix
+    to avoid name collisions.  Also we will create another form that controls
+    if the accounts are enabled.
+    """
+    form_classes = {
+        'kaltura': KalturaAccountForm,
+        'brightcove': BrightcoveAccountForm,
+    }
+    def __init__(self, team, data=None):
+        self.is_bound = data is not None
+        existing_accounts = dict(
+            (name, form_class.get_account(team))
+            for (name, form_class) in self.form_classes.items())
+
+        enabled_accounts = self.make_enabled_accounts(existing_accounts, data)
+        # trigger a full clean on enabled_accounts since we want to use
+        # it's cleaned_data below
+        enabled_accounts.is_valid()
+        self.forms = {
+            'enabled_accounts': enabled_accounts,
+        }
+
+        for name, form_class in self.form_classes.items():
+            # if the account is enabled, then we pass it the POST data,
+            # otherwise we leave it unbound
+            if self.is_bound and enabled_accounts.cleaned_data[name]:
+                form_data = data
+            else:
+                form_data = None
+            self.forms[name] = form_class(team, form_data,
+                                          instance=existing_accounts[name],
+                                          prefix=name)
+
+    def make_enabled_accounts(self, existing_accounts, data):
+        fields = {}
+        for form_name in self.form_classes:
+            fields[form_name] = forms.BooleanField(
+                required=False, label=_('Enabled'),
+                initial=existing_accounts[form_name] is not None)
+        EnabledAccountsForm = type('EnabledAccountsForm', (forms.Form,),
+                                   fields)
+        return EnabledAccountsForm(data, prefix='enabled_accounts')
+
+    def account_forms(self):
+        for form_name, form in self.forms.items():
+            if form_name != 'enabled_accounts':
+                yield form_name, form
+
+    def account_enabled(self, form_name):
+        return self['enabled_accounts'].cleaned_data.get(form_name)
+
+    def is_valid(self):
+        if not self.is_bound:
+            return False
+        return all(form.is_valid()
+                   for (form_name, form) in self.account_forms()
+                   if self.account_enabled(form_name))
+
+    def save(self):
+        if not self.is_valid():
+            raise ValueError("form not valid")
+        for form_name, form in self.account_forms():
+            if self.account_enabled(form_name):
+                form.save()
+            else:
+                form.delete_account()
+
+    def keys(self):
+        return self.forms.keys()
+
+    def __getitem__(self, form_name):
+        return self.forms[form_name]
+
